Move some logic from NoteHeader to Note
